Creating a new route for login
