+++ server/routes/FindGamesList.js
@@ -1,0 +1,1 @@
+// Getting a List of games according to User Params
